--- src/webComponents/helpers/methodsMock.ts
@@ -1,28 +1,0 @@
-export default [
-  {
-    id: 0,
-    name: 'Environment',
-    children: [
-      {
-        id: 92,
-        name: 'Product',
-        children: [
-          {
-            id: 929,
-            name: 'Service',
-            children: [
-              {
-                id: 20,
-                name: 'API 1',
-              },
-              {
-                id: 24,
-                name: 'API 2',
-              },
-            ],
-          },
-        ],
-      },
-    ],
-  },
-];
